requirements.txt
djangorestframework
asgiref==3.3.4
Django==3.2.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
